doc: (Switch) split demos
