--- src/components/switch/demos/index.less
@@ -1,3 +1,0 @@
-.my-switch {
-  --checked-color: #00b578;
-}
